Here is a version that produces the right values of max(b^2) and mean(b^2) in the rings1 directory. I've also checked in the Makefile.local and cparam, so you should be able to reproduce the following:
 reduced viscous force
 diagnostics, nmax_count, nsum_count= 1 1
 fsum=   7.2846045E+03
 fmax=   4.1641746
 entered reduce
 entered reduce2, nw, ncpus= 32768 1
 iname,itype_name(iname)= 1 -1
 end of pde
 timestep ok
 timestep ok2
 it,t,bmax                          brms
 0   5.3999852E-02   4.1641746   0.2223085

 after prints
*** Arithmetic exception:  - aborting

So the values 4.1641746   0.2223085 are correct (checked with idl),
but then I don't know why it crashes. In principle, one should be able
to interchange the entries in print.in, but then it crashes earlier.
All it does it (theoretically) interchange the sequence of printing...
I better stop here. I'll call you first.