1. Added lreset argument in rprint_list (and related) routines to allow    changing the output format during the run. 2. moved init_hydro from register.f90 into hydro.f90 3. it,tdiagnos,dt no longer by default: need to be given in print.in 4. ampl removed from several routines (treated as in magnetic.f90 as    part of the relevant module only and need to be read there)
I think this is about it; there may be other smaller things.
I did test spher1 and vconvect1 for i/o consistency, but in spher1
I also noticed that the length of the timestep changed (with my new
version relative to your previously checked in version). Could that
be because of some missing ampl somewhere, or it the random noise
just different somehow?

We should think about ways to check that certain new revisions are
still giving consistent numbers.