Restructured IO of init/run parameters: - all parameters are now read in as namelists - introduced new module Param_IO - renamed cread/cprint to read_runpars/print_runpars and moved to Param_IO - Added nl2idl, a Perl script that reads a namelist and writes and IDL file   generating a structure `par' with a slot for each variable from the   namelist.
The IDL stuff doesn't work smoothly as yet: sometimes you need to run
start.pro and r.pro twice before it understands tmp/param.pro and
tmp/param2.pro, although it is not clear at all why this happens.
  Also, I am not sure the current philosophy is good: do we really want to
use structures (in order to avoid cluttering the namespace with unimportant
parameters)? If yes, we will probably want `par' for param.nml and `par2'
for param2.nml (currently it is all `par', so nl2idl needs to be improved)
and we probably don't want to copy that many parameters as currently (like
lentropy=par.lentropy, etc), but rather work with par.XY where we need
parameter XY. If no I also need to modify nl2idl.
  At least vconvect1 seems to work, although there might still be something
wrong: I don't get the same output as before, but it might in fact be that the
old output was wrong and I have fixed somthing along the way.