Removed old errmesg mechanism for flagging errors in boundcond.
According to what Nils said, I have moved the call to boundconds() after
that to finalise_isendrcv_bdry().

However, we still have a problem with the corners, i.e. the edges in 3d.
The x-y edges for example are not OK, which is difficult to identify, but I
see it when I compare nprocy=1 and nprocy=2. Just like in the case of
communication, we must treat the edges extra if we combine boundconds
(in the x direction) and communication (in the y direction).
  I am too tired for doing this now and possibly too lazy to start this
messy work tomorrow.

I think one thing that would help is to do communication and (then) boundary
conditions in x first, then doe comm. + boundconds in y, then in z. But I
don't think we would want this, since currently the communication takes
place simultaneously for all three directions, which would no longer be
the case.