> I can at least confirm that  > brandenb@mhd1.st-and.ac.uk:/home/brandenb/f90/pencil_modular/runs/forced/hel1_4procs>  > brandenb@mhd1.st-and.ac.uk:/home/brandenb/f90/pencil_modular/runs/forced/hel1>  > gives the same results!
If I do five time steps of hel1_4procs with 4 CPUs, I get what you find at
  http://www.kis.uni-freiburg.de/~dobler/MHD-code/communication.html
Clear artefacts on the z-CPU boundary.

I forgot that the current forcing routine uses a "halo" with a default
scale height of pi. I put now the default value in the fortran code
to zero to turn off the halo (will be checked in soon). Here the
current data file (for future reference).