Fixed problem with ran0() [which was trying to modify the constant value 10]. Also added the routine ran() from NR for F90 that does the same job.
The problem with these generators is that they will generate the same
sequence of numbers on each CPU. While this is what we want for stochastic
forcing, it is not OK for random initial conditions.
If I remember correctly, I have circumvented this problem somehow in the
initialisation for the builtin generator.