Wolfgang et al,
1) Go get a coffee
2) Start here...
-----------

Viscosity calculations for entropy and momentum equation seperated into
an new module called Viscosity. (As discussed with Axel)

Default (in Makefile) implementation visc_const.f90 implements present
zero and constant viscosities, infact is a direct rip.

New variant visc_shock.f90 will calculate shock viscosity which is expected to
have a noticable overhead in the pencil code.

NEW viscosity_run_pars namelist contains only nu and ivisc (as in previous
code).  They have not yet been removed from their previous location to
temporarily allow backward compatibility.  This is possible as
viscosity_run_pars is the last namelist in run.in and hence any values
set by hydro_run_pars will be over written iff they are defined in
viscosity_run_pars.  Ultimately they should be removed from hydro.

There is currently  not a 'noviscosity.f90' module as this functionallity is
available in the same way as it was before i.e. by setting nu=0. (the default)
This means, to stay within the current framework the lviscosity boolean is
ALWAYS true...
********************

TO RESOLVE:
There are 2 places in the code (marked by my initials - grep ajwm *.f90)
where a condition exists which was if (lentropy .or. ivisc='nu-const')
To keep the viscosity variables contained within the viscosity module
this has been changed to if (lentropy .or. lviscosity) which is effectively
if lentropy.

The two places are the calculation of [grad lnrho] and of [sij] so there is
an noticable overhead if they are calculated unnecessarily.  But this condition
may grow more complex.  Options are:
   a) give up on data hiding declare ivisc in cdata again
   b) produce logicals eg [need_sij] etc. ored with .true. by any modules
      that want them, on startup

This issue also affects [nu] which is currently used by two cases in forcing
so has been made global again temporarily (i.e. declared in cdata.f90).

ivisc should be renamed, it is no longer an integer.

*******************
OTHER CHANGES

Minor change removing the occurance of DD in cdata.f90 and correction of
comment in pscalar.f90 (DD -> pscalar_diff)