Added a grad ... in momentum eq!
A note on the zero bulk viscosity mentioned in the manual:
According to one of my "old" compendiums  one can show using the kinetic theory
of gasses that the bulk viscosity of a  monoatomic gass is zero. But according
to (Truesell, J. Rat. Mech Anal V. 1. (1952)) this a result of the assumptions
underling the kinetic theory of gasses. Experiments show that it close to zero
for monoatomic gasses but large for more complex fluids. It is usually
safe to drop it because the deviator of the strain is much smaller than the
strainless tensor not because it is small :-)