Added and extra control file, RERUN
Very naughty but allows one to rerun the pencil-code without resubmitting.
Useful for debugging but it could be used irresponsibly.  If the file is
empty the code simply runs again.  Alternatively the file may contain
a path which will be changed to before the code is RERUN.

(I'll put this in the manual it is not considered too naughty!...)

eg.  to simply RERUN
     touch RERUN         <--- Tells the script to restart when
                              the pencil code exits
     touch STOP          <--- Tells the pencil-code to exit

or perhaps something like,
     echo '/home/pencil-code/runs/mynaughtyrun' > RERUN
     touch STOP
to rerun using the parameters etc in, mynaughtyrun!

RERUN is not presently supported by start_run.csh.  There is also the
possibility of having a RESTART file too...