Split up the condition:
   if ($?JOB_ID && -e $HOME/.score/ndfile.$JOB_ID) then

csh does not appear to be completely C-like in its evaluation of
conditionals.  C would not make any attempt to evaluate the right hand
side of the && if the left hand side were found to be false.

csh appears to, at least, expand the variables in the right hand side
even when the left hand side is false.  Hence an error as $JOB_ID doesn't
exist.  Only showed up when running interactively on copson.