Trying to figure out why compilation sits on register.f90 for so long.
Idea:
Strange that the module with the least code takes the most time to compile.
It does however have the most modules as dependancies.  Perhaps compiler is
spending time looking in all possible places for external references and
telling it explicitly what to get from where shortcuts this process.

Changes:
Be more explicit with the use statements... Seems to slightly speed up
compilation of the register.f90.

Probably better practice anyhow.

Checked in so I can switch between revisions for testing a little easier.