Following discussion with Axel.  Checked in my code for allowing 'non-standard'/special customizations to the Pencil-Code.
This serves two purposes...
   1) It allows the addition of new physics etc. which would not sit well
      with the main code. eg. special damping terms, approximations for
      higher dimensional effects if 1D, 2D runs etc.

   2) It allows for the development of additional physics code which might
      later, i.e. once tested and shown to be useful, be integrated in to
      the code. eg. geodynamo potentials, spherical boundaries

It provides a set of standard hooks into the numerical parts of the code
to allow somebody to add terms to the various PDEs
(currently: hydro, density, entropy and magnetic only) by providing
subroutine calls with standard interfaces as hooks.

I have not yet enumerated all the possibilities so am submitting this as
a request for comments.

To use this a user should copy nospecial.f90 to some other name (i.e. using it
as a template).  Fill out the function prototypes the require to use. And ensure
that they correct the register_special and initialize_special routines.
nospecial.f90 is 'over commented' in an attempt to guide the user through it's usage.

There is also the ability to have extra arbitrary coupled PDEs and new
initial conditions but this has not been implemented as yet...

I was trying to avoid multiple modules but that may be necessary:
  SPECIAL_TERMS     - for extra terms on the standard PDEs
  SPECIAL_EQUATIONS - for new coupled PDEs
  SPECIAL_BOUNDARY  - for special boundary conditions
  SPECIAL_INITIAL   - eat up the INITCOND_SPEC thing!

---------------------------------------

Axel chose the name SPECIAL over CUSTOM and PLUGIN.  Feel free to comment
on this and I shall feel free to ignore you :-P

----------------------------------------
Wolf,
The ability to add mvar & maux variables would require a user to be able to
declare what contributions their module would make.

Presently the generation of cparam.inc requires some hard coded numbers in
the script.  I was thinking making all the modules (not justn this special stuff)
have a standard comment somewhere near the top like:

! MVAR CONTRIBUTION 1
! MAUX CONTRIBUTION 0

Which could found with a perl regexp match from all the source files
add added to give the values for cparam.inc

See the top of nospecial.f90 What do you think?
Your perl abilities are far in excess of mine... Could you throw a
script together to do this? It seems like it should be about
5/6 lines of code.

eg.  I'd call:

mk-cparam $(DENSITY).f90 $(ENTROPY).f90 .... etc
  in the Makefile

This would remove the hard coded info from mk-cparam and put it with the
rest of the information about each module... in the source file!