Modified method by which mkcparam works...
mkcparam now takes a list of all the source files (to be used) as parameters
and parses each one looking for lines of the form:

! MVAR CONTRIBUTION xxx

and

! MAUX CONTRIBUTION yyy

where xxx and yyy are the respective contributions to the size of the f
array (and df array for xxx). Hence there is no code dependant information
hidden in mkcparam anymore.


Added decleration lines to all the physics source files. And modified the
Makefile accordingly.

Had to a a new $(physics_src) variable in the Makefile, unless there are any
better suggestions.

Haven't finished auto-test, but will checkin and complete on a faster machine!