Modified Makefile to always use a specifically named 'makefile.adapted' instead of just 'makefile'.
This is to allow less competent machines, like Macs with HFS+ file systems,
to correctly process the adapted makefile.

HFS+ is case insensitive to all intent and purpose so screws up at some point
if the Makefile vs. makefile cleverness is used.

This should still work on all other architectures and is more explicit than the
magical makefile.

    -- tOnY checking in from Dave's inferior PowerBook,
           but which succesfully runs the Pencil-Code