Irrelevant changes in docs.
In the first place, I want to check the new command line in
CVSROOT/loginfo. This Email should have a `Reply-To:' field and when you
tell your mailer to reply, the answer should automatically be deirected to
dobler@nordita.dk.
  I think all of the CVS users also have an account at Nordita, right? So
from now on, replying to a CVS-generated email sends mail to the right guy
instead of Axel.

Keep in mind that this requires you should check or forward your nNordita
Email.