Added inner looping over multiple processors per load for varfile distribution.
Set one_local_disk=0 for copson

Set nprocpernode = 1 by default.

Wolf/Axel: Does this still work on Horseshoe etc?
Dave: It works on copson...
   qsub -pe score XX -masterq copson.q SOMETHING.csh

where XX is the number of slots = number of nodes + 1
      and SOMETHING is start | run | start_run

For some reason SGE is running the batch script on the first node instead of the masterq
hence it can't find the scout executables...  I should maybe ask Keith why this is not the
default behaviour on the cluster.