Added a routine parse_shell to sub.f90 to parse strings containing $IDENTIFIER where IDENTIFIER can be any combination of alphanumerics hyphens and underscores, replacing occurances with the value of the appropriate environment variable.
Presently this is only called by the get_snapdir routine
in param_io.f90.  It allows of the file
   directory_snap
to contain expandable environment variables.

This allow things like
  setenv SCRATCH_DIR '$TMPDIR'
as in the new copson.st-and.ac.uk section of getconf.csh.
This does not break the present file distribution either.
(I don't think!)

---------------------------
The new code is dependant upon the GETENV function call.
I considered the portability of this an believe most
and hopefully all the current compilers support it some how.
I started to try and catalogue them, please add on:

 http://www.students.ncl.ac.uk/a.j.mee/archives/000097.php

The code auto-tests for me but if it doesn't work
(i.e. there is no GETENV function to be found)
Commenting out line 2536 of sub.f90 - where GETENV is called -
should solve the problem.

There is also an interface definition at the top of sub.f90,
but this should be ok.

----------------------------

Added the -Vaxlib (include portability function library)
option to all the ifc command lines in Makefile.src

---------------------------