Finally fixed copson configuration to use new distributed scratch space.
Currently set to use gm2 PE which will be renamed to gm this evening.

Requires bc to be coppied to your path as it doesn't exist on all the copson
nodes (I've just put a copy in $PENCIL_CODE/bin for now)

Perhaps we could find a work around for bc, command-line perl perhaps?