Just bug hunting and minor optimizations.
The temperature is now sent to dndt_dt and on to get_growthtime. Added
second continuity equation (for rhod) when both rhod and nd used. Also when the
lowest mass dust bin underflows, the mass is transferred to the passive scalar
measuring concentration of condensable material in the gas.

Replaced ldustformation with ldustnucleation. Abandoned work on dust
nucleation. This is tricky stuff, and not even necessary at the moment (other
dust particles will work as condensation nuclei for the ice).