Replace 'dummy' specification with 'UNKNOWN' and actually read data from variables tacked on the end of the farray.
I.e. if one adds a variable to the farray and doesn't add it to varcontent.pro
it will still be loaded in IDL but into a variable called UNKNOWN. And it's
stats will be displayed.

If more than one UNKNOWN variable occurs in a var file the last one will be read
and stored in UNKNOWN.  This should encourage one to define this variable
if it's a keeper or just use it temporarily as UNKNOWN if it is simply for debugging
purposes.