added new possibilty of adding subsequent initial condition -- so far only for density and entropy.
initlnrho and initss are now declared as character arrays. This makes it
possible to say in start.in for example

  initlnrho='const_lnrho','blob'
  lnrho_const=1
  ampllnrho=0.1
  radius_lnrho=1

which first sets lnrho=1 everywhere and adds a blob afterwards. Currently up to
4 subsequent initial conditions may be specified. specifying just one initial
condition still works as well. I have tested this on an Intel and an IBM
machine. Hope this doesn't brake anything for people using SGI or other
machines.