Added xH2 which measures fraction of molecular hydrogen to the total number of hydrogen atoms.
The total number of hydrogen atoms is nHtot = nH + 2 nH2. This means that when
nH2 = 1 and nH = 0, then xH2 = 0.5. The code gives a warning if someone (me!)
should accidentally set xH2 = 1.0

This passed the auto-test. I hope it does not disturb something else.

Can be easily transferred to ionization.f90 and noionization.f90 if anyone
needs it.

Also the subroutine getmu has the optional input mumol which is the true
mean molecular weight, not the annoying mu1yHxHe.