Redistribute dust grains in correct bins before boundary conditions are communicated with MPI. This is done from pde then. I am not aware of any 'official' place to put pre-communication fiddling with the f array.
In dndmd_dt all ice is evaporated immidiately when evaporation sets in. The
problem before was that ice evaporation is proportional to dust grain surface
area and does not speed down when the ice is running out, so ice masses became
negative.

With lbogusdustder (which really needs renaming) all continuity equation ice
transfer is transformed to a change in the passive scalar (water gas).