Host specific variable notserial_procN controls whether processors are numbered as
  nq2: proc0,proc4
  nq4: proc1,proc5
  nq5: proc2,proc6
  nq6: proc3,proc7

or

  nq2: proc0,proc1
  nq4: proc2,proc3
  nq5: proc4,proc5
  nq6: proc6,proc7

The latter is almost always the case, but the Nordita cluster uses the first.
This took some rewriting of the copy of var.dat to all the different nodes in
run.csh. The problem before was that wrong var.dat files were sent to the nodes.
Since this error was extremely difficult to find, I set as default output of
the copy command (this shows which nodes get what and where it is put).

In getconf.csh $nnodes is now calculated at the number of elements of
$nodelist. This allows for always printing out the number of CPU(s) per node.