Fixed some more brackets.
Moved varcontent.pro to functionalized version in files/pc_varcontent.pro
which returns the varcontent array as its result.

This version can be used by the pc_read_* routines, and is also now
used by r.pro, rall.pro and rallxy.pro so that this information is only
in one place.