"Thou shalt not fiddle with the f array, or a grave error will appear before you. Yet this shall only occur a full day and a full night after you start the simulation as this will make the pain worse. In stead you must change the accelerations such that the result of your desire is indeed obtained in the next time-step."
When the stopping time is comparable to or shorter than the time-step, the
acceleration is now set such that the dust species will obtain the same
velocity as the gas (but with terminal velocity in the z-direction). The old
fiddling scheme was the (only?) source of assymetries in the dust runs.