Put in some documentation in the head of hydro_fixed, density_fixed and entropy_fixed.
Here is an exact copy of the original commit message:

"Added files hydro_fixed, density_fixed and entropy_fixed. These put the
relevant variables as auxiliary variables. Input parameters are still read in,
but there is no dynamical change to the variables."

This still applies. It is useful for considering dust dynamics without spending
time on the gas dynamics. Example: Consider a gas vortex that concentrates
dust. The dust concentration depends on the gas density and the gas velocity
field. Thus the gas variables can not be ignored, but the dynamical evolution
can be ignored for the purpose of wanting only to see what happens to the dust.

These modules are likely to become outdated if they are not frequently used or
auto-tested.