Several improvments for non-equidistant grids. Now possible in all 3 dimensions + proper time stepping formulae (see last check-in). Most of the grid setup is now done in grid.f90.
To use a non-equidistant grid in the x-direction

- set grid_func(1)='sinh'
- set coeff_grid(1) to specify the ``shape'' of the grid funtion
- set xyz_star(1) to specify the inflection point of the grid function

sinh as the grid funtion gives -- dependend on coeff_grid -- a linearish
profile near the inflection point and a rather exponential profile away from it.

See the manual and comments in the code for some details.

Diagnostic routines such as sum_mn_name have not been modified to give correct
results in the non-equidistant case yet. Also the IDL routines for calculating
derivatives need to be adjusted.