The line with
  if (ndustspec == 1) sdust=''

in dustdensity.f90 was commented out after Tobi's latest changes to the
time-step. The log message says nothing about changing to above statement, so I
assume it was a mistake. This caused the rhodm diagnostic output in
dust-vortex to be zero, so new reference data with working rhodm is checked in
as well.