Fixed problem with setting qshear in run.in. This did not set Sshear accordingly, since Sshear was already defined from earlier value of qshear (set in start.in or in run.in before reload). The potential problem was noted by Tony in a remark in initialize_shear.
Now Sshear is set to Sshear = -qshear*Omega whenever qshear is not zero. This
should be alright. I searched the entire repository for the presence of Sshear
in input files, but nothing there, so this change will not affect anyone
currently.

Small concern: does it make any sense to set Sshear differently from
-qshear*Omega? Because this is no longer possible.