Fixed bug for when LENGTH wasn't specified.
Use angles measured CCW from the +ve x-axis as per other IDL/computing routines.
And some comments describing this.

Thought it was already so but it turns out the component based code was,
internally using angles calculated CW from the +ve y-axis (as per a compass)

The former has been chosen for consistency with other computing tasks.

Of course one can achieve the former with:
   wdvelovect,U,(!pi/0.5)-V,/polar

(Thanks to Andy for pointing out that this wasn't working as I intended.)