Include routines relevant to the pc_radio radio telescope observation code for performing gaussian smoothing by various methods.
From header of README:
--------------------------------------------------------------------------
THE IDL ASTRONOMY USER'S LIBRARY (March 2004)

The IDL Astronomy Users Library is a central repository for general purpose
astronomy procedures written in IDL.    The library currently contains
500+ procedures including general FITS I/O, interfaces with STSDAS and IRAF,
astronomical utilities, and photometry and spectral analysis procedures.
The library is not meant to be an integrated package, but rather is a
collection of procedures  from which users can pick and choose (and possibly
modify) for their own use. Submitted procedures are given a cursory testing,
but are basically stored in the library as submitted.     The IDL
Astronomy User's Library was funded through November 2000 under the NASA
Astrophysics Data program.

The IDL Astronomy Library is accessible via the World Wide Web home page
http://idlastro.gsfc.nasa.gov/homepage.html.     The default version of the
Library requires IDL V5.3, although earlier versions are available for
V4.0-V5.0, V5.1, and V5.2.