Cleaned up Slices module. Now both declarations and slice writing occurs in the order
  code variables -- auxialiary variables -- derived variables

This makes the view much better. Please keep tidy from now on :o).

Implemented slices of dust-to-gas mass ratio (epsd). This requires knowledge of
grain mass that only the dust modules have, so the grain mass slices are
calculated in Dustdensity.

Personally I find it very dissatisfactory that so many slices must be written
outside the Slices module. I guess the solution is to pencilise Slices, so
that's an idea for a future project.