Needed to check in quickly at 4am this morning to set something running somewhere else and couldn't find any machine that on which vi highlighting worked!  Hoped someone would tell me.  All is now restored.
The fault is in fortran.vim v8.2 (possibly all < v8.6).

For others using vi and finding huge chunks of text following ""
in fortran files appear red (as if the last " was missing):

Placed a copy of the fortran.vim v8.6 in ~/.vim/syntax
and this appears to fix the problem on machines with
vim >= v6.0.

fortran.vim v8.6 can be found installed on machines with vim v6.3 in:
  /usr/share/vim/vim63/syntax/fortran.vim

One may wish copy the following files from a newer vim distribution to appropriate ~/.vim paths too.
  /usr/share/vim/vim63/ftplugin/fortran.vim
  /usr/share/vim/vim63/indent/fortran.vim


----------------------------------------------------------------------
>  > "" seems to break vi syntax highlighting on some systems...
>
> If it's only some systems, it canm obviously be fixed..
>
> > If you think this is crazy and can fix vi instead please let me know.
>
> I basically do, although if it really can't be fixed on Vim's side, we'll
> have to live with it.
>
> Which version of vim is that? And which version of fortran.vim?
> I don't see this problem with
>  vim 6.3
>  fortran.vim 0.86 (can send it if you want)
>
> W o l f g a n g