Replacement interstellar auto-test.
If anything's gonna break MPI symmetries... This should do it.
32^3 points on 4 processors!
Modelling a very low-res 250^3 pc chunk of the ISM, with a single
SNe near the box centre, but well into the 4th processor.

Shocking really.

Nice and quick so many more timesteps (30).  I think this is now a
very strong test.  With the timestep bug of the previous days
it dies catastrophically after just 2 timesteps.