Must change nodelist for one_local_disc before directories are made on the scratch disc. This also only requires the work of one node.
For some machines with common scratch disc (such as RZG in Garching) there is
only rsh access to the gate node. For this purpose the variable masternode can
be set in the machine dependent part. The master node is then later set as
nodelist (if $masternode is defined), instead of the first node of the nodelist
being set as the nodelist (default, when $masternode is not defined).