Removed old (size based) method for copying snapshots.
Added a lot of comments both in the program and in the output. If this is too
much for some people's taste, we could consider having different levels of
debug (as in the code).

I attempted to not change anything in the actual copying method, so everything
should be okay, but please be aware of this new version of copy-snapshots if
you run into trouble.