Experimental new version of copy-snapshots. Relies on copy-me.list files present in the proc* directories at the "global" disc (ie. not the scratch disc). These files are currently only written by my local version of the code.
The files copy-me.list are compared with the files already-copied.list. All
differences are immidiately moved from the local scratch discs, and the
filename is then appended to already-copied.list.

The biggest concern right now is the effect of having all the processors write
to the global disc everytime a snapshot is written. Even though the
copy-me.list files are small, this could potentially either slow down the code
or cause the filesystem to go crazy (if several hundred processors are writing
files at the same time). I am going to test soon on Horseshoe.

Later improvements would include looking for copy-me-aux.list, a file written
by the user. For example

  echo var.dat > data/proc*/copy-me-aux.list

would cause copy-snapshots to move var.dat from all nodes (and not register the
filename as copied).