Implemented nonlogarithmic dust density as an option (set by the logical ldensity_nolog). Only requires a few logical switches in dlnrho_dt.
Also implemented hyperdiffusion on gas density (which is why I needed the unlog
dust density in the first case).

lnrho, rho and rho1 are now all calculated in calc_some_vars, and then sent to
the appropriate subroutines from pde. This would save a lot of calculations of
rho and lnrho, so think it was worth the effort to calculate lnrho, rho and
rho1 only once.

Unlog gas density may not work too well for exotic modules such as radiation
and cosmic ray.

The auto-test went well, but still there may be some cases that were not caught
by the tests...