Catch USR1, USR2 signals and do controlled job shutdown when they arrive.
With Gridengine (or other queuing systems that notify a job before
shutting it down), submit your job with
  qsub -notify run.sh
and perl_monitor will touch RELOAD when the job's time is over.
If notification is not configured by the administrator, use
  qsub -h_rt=48:00:00 -s_rt=47:55:00 run.sh
which will give the code 5 minutes time to write the data and finish.

Thus I don't need to estimate a good number of time steps in
advance. PBS unfortunately does not seem to support this.