nl2idl now uses new Perl module.
PLEASE PORT THIS TO eos AND TRY IT OUT (ON BOTH BRANCHES)!

The Perl module will form the basis of a script that fixes the
start.in/run.in parameters automagically for the transition
pre-eos -> post-eos.
I need any feedback to make sure that the module works correctly (if it
doesn't, don't forget to send me the namelist files).