Changed the calculation of the index of the processor that should receive a migrating particle. Now it is done by succesively subtracting Lxyz_loc to the particle position and seeing when the result crosses the box boundary.
The old method of calculating the index directly failed sometimes (rarely), but
even those few failures turned out to be too annoying in the end.

The new method seems to be around 4% faster, but this may change if many
particles cross box boundaries, because then the receiving processor may not be
next to the sending, and it requires then many more iterations to find the
index.