Put a lot of comments in the code.
Fixed so that stats are written also for trimmed variables in runs with less
than three dimensions. Did this by sending the information about whether the
data is trimmed or not into pc_is_scalarfield and pc_is_vectorfield. This is
simple, but it would of course be nice to have a program that can determine
uniquely from an array whether the array is trimmed or not.