Continued Tony's work on auxiliary variables with boundary conditions. These are declared as COMMUNICATED AUXILIARIES in the header of a module.
The idea is that they lie just after the mvar physical variables. The sum of
communicated auxiliaries and mvar is mcom. So simply changed all boundary
conditions and boundary communications to consider the first mcom variables
instead of the first mvar. This also works for shear.

An example of a communicated auxiliary is the shock viscosity. Since it is not
constant, the diffusion terms involves the gradient of the shock viscosity, so
ghost zones must be set.