Put boundary conditions on uu after communication. Put x-boundary conditions on shock after it has been calculated.
This should be the proper way to make shock (shear) periodic. Built for comfort
rather than speed, at the moment uu is communicated twice and boundary
conditions are applied twice. So this can probably be optimised somehow.

Tony, this breaks interstellar auto-test. Can you check if it looks okay? I
think this is a proper correction of the shock boundaries.