Removed local_isothermal subroutine from the EquationOfState module. It was too specific to be in such a general stuff like eos. But there is a problem and I would be glad if someone could help on fixing.
I tried to put local_isothermal in the Planet module, and then call it from
noentropy using

Use Planet, only :: local_isothermal

but it says it can't find the Planet module library. So, I had to move the subroutine to noentropy.f90 instead. It's not a good location, but it's where it is called (I use it to allow the sound speed to vary although the entropy equation is not solved)