density.f90 -> purged mass_source_spline and added total mass of the disk as monitored quantity
sub.f90 -> added a subroutine sum_lim_mn to integrate the density over the predefined limits r_int and r_ext.

Question for Axel and Wolfgang: the mass is conserved when using nolog_density, but IT IS NOT CONSERVED WHEN SOLVING FOR LNRHO. Why, oh gods of numerical hydrodynamics, why???

By the way, nolog_density and non-smoothed gravity solved by mass conservation problem in the inner boundary that I though was due to poor resolution. I don't even need freezing anymore. Apparently, it was not, and a cartesian grid works quite well in this highly curved rapidly rotating inner zone of the disk. Very good, but also unclear why.