Place particles for streaming initial condition according to density distribution function. Should be made more general so that all initial conditions can do this.
Setting e.g. n(x,z) = n0 + A*cos(kx*x)*coz(kz*z) is not simple for particles.
One can however solve the equations

  int_{x0}^{x}[n(x,z)]*dx / int_{x0}^{x1}[n(x,z)]*dx = r
  int_{z0}^{z}[n(x,z)]*dz / int_{z0}^{z1}[n(x,z)]*dz = p

where r and p are random numbers between 0 and 1. This involves most of the
time a numerical inversion of the equations so that x(r) and z(p) are found.