Added new module that directly evolves logarithmic temperature instead of entropy. As the name suggests, it uses an equation of state that includes hydrogen ionization. Currently, it only gives correct results for 0-D runs.
The new EOS module is a marginally modified version of noeos.f90 right now. To
make it useful, all the EOS modules will have to ``pencilized''.

In the 0-D test run that I will commit after this, the speed difference to the
entropy version of the same run is dramatic. Further advantages are greatly
simplified expressions for the thermodynamic quantities and much better
numerical stability.