1. added the atomic mass unit (which is usually used for Rgas). But to make Tobi's test runs work, I commented it out again. But we need to think about it.
2. In entropy, I included (but commented out again) the possibility
to print T in Kelvin.

3. In eos_temperature_ionization, I've added gamma, but I really think
we should enable this again for reading namelists etc to steer this.
gamma is used in interstellar, for exampe. Maybe it shouldn't.

4. In interstellar, I captured some cases where temperature wants
to write into ilnTT, not iss, of course.

When I uncomment some of the above, I am able to reproduce the cooling
test more or less correcty with temperature_ionize (with noionize!).