Made the eos module pencil_case aware to allow better integration of the idealgas/ionization etc. code.
eos_idealgas is now explicit in the various twists (isothermal, isentropic)
and is generalised to work with any two thermodynamic variables as the
"evolved" or base quantities (ie. those in the farray or held constant)
At present lnrho/rho must be one of the two and the lnTT,ss and cs2 cases
are handled.  The alteration to make any two thermodynamic quantities
work is trivial eg, if you wanted to work with pressure and squared sound
speed (if you are that kind of sick puppy). An example of the flexibility
is given by handling Wlad's neat "locally isothermal" equation of state
that was previously bodged into the planet module.