Sort particles so that they come in order of ascending imn. Thus they can be treated spatially together with the grid variables.
The sorting is done using a simple straight insertion algorithm. The first sort
takes literally ages, but once the array is almost in order, then the sorting
does not take very long. And there's loads of possible better performance when
the particles are treated in the m, n loop.

Advanced programming hint of the day: use \cp for backup, never \rm...