Implemented a shell sort algorithm for sorting particles. It is much faster once Keplerian shear is present (or just a high particle velocity rms).
The result now is that the code is faster when the particles move slow, whereas
with Keplerian shear or just a high general velocity, it can be a bit (30%)
slower. Still the gain in better code structure (particles can use pencils!) is
immense. And the sorting can probably be optimized some more.