Added support for the IBM Linux cluster in Copenhagen (http://www.dcsc.ku.dk/). This is a dual processor, dual core AMD opteron machine.
The cluster uses IBM loadleveler as queueing system, and since llsubmit (the
equivalent of qsub) doesn't seem to accept any useful command line arguments,
I wrote a wrapper script (llsub) that prepends start.csh and run.csh with
Loadleveler directives. Job submission looks like this:

  llsub -n 8 -w 01:00:00 run.csh

llsub is my first perl script and I'd like to ask any perl expert to take a
look and improve it.

Not related to this, the code that creates the directory structure (procdirs
and subdirs) on local scratch discs assumes that everybody uses csh, which is
of course not the case... I added a fix for this which is currently commented
out, but I'd like to ask everybody to test it. (getconf.csh)

Also, don't try to copy timeavg.dat if it doesn't exist (run.csh).